break components
